+++ features/booking/entity.go
@@ -10,8 +10,8 @@ type Core struct {
 	CampID        uint
 	Title         string
 	Image         string
-	Latitude      float64
-	Longitude     float64
+	Latitude      complex64
+	Longitude     complex64
 	Address       string
 	City          string
 	CampPrice     int
